Move models to a more descriptive file
