+++ app/models/db/MetricsFilters.scala
@@ -2,11 +2,8 @@ package models.db
 
 import com.github.tototoshi.slick.PostgresJodaSupport._
 import com.gu.editorialproductionmetricsmodels.models.{OriginatingSystem, ProductionOffice}
-import io.circe.Encoder
-import io.circe.generic.semiauto.deriveEncoder
-import io.circe.syntax._
 import models.db.Schema.DBMetric
-import org.joda.time.{DateTime, DateTimeZone}
+import org.joda.time.DateTime
 import org.joda.time.format.ISODateTimeFormat
 import slick.jdbc.PostgresProfile.api._
 import slick.lifted.Rep
@@ -105,38 +102,3 @@ object MetricsFilters {
       filters.newspaperBook.fold(TrueOptCol)(nb => metric.newspaperBook.toLowerCase === nb.toLowerCase)
   }
 }
-
-case class CountResponse(date: DateTime, count: Int)
-object CountResponse {
-  private val datePattern = "yyyy-MM-dd'T'HH:mm:ss'Z'"
-
-  implicit val timeEncoder = new Encoder[DateTime] {
-    def apply(d: DateTime) = d.withZone(DateTimeZone.UTC).toString(datePattern).asJson
-  }
-  implicit val metricEncoder: Encoder[Metric] = deriveEncoder
-}
-
-case class ForkResponse(issueDate: DateTime, timeToPublication: Int)
-
-object ForkResponse {
-
-  def convertToForkResponse(pair : (Option[Int], DateTime)) = {
-
-    val (timeToPublicationOpt, forkTime) = pair
-
-    timeToPublicationOpt.map(timeToPublication => ForkResponse(forkTime, timeToPublication))
-  }
-}
-
-case class WordCountResponse(path: Option[String], wordCount: Option[Int], commissionedWordLength: Option[Int])
-object WordCountResponse {
-  implicit val wordCountResponseEncoder: Encoder[WordCountResponse] = deriveEncoder[WordCountResponse]
-}
-
-case class GroupedWordCount(countRange: (Int, Option[Int]), count: Int)
-
-object GroupedWordCount {
-  implicit val groupedWordCountApiResponseEncoder: Encoder[GroupedWordCount] =
-    deriveEncoder[GroupedWordCount]
-}
-
